backend/requirements.txt
Flask
firebase-admin
openai
python-dotenv
